requirements-dev.txt
# E2I Causal Analytics - Development Requirements (Locked)
# Generated from working venv on 2025-12-22
# Python 3.12+ compatible
#
# Includes all production dependencies + dev tools
# Install with: pip install -r requirements-dev.txt
# Or install from pyproject.toml: pip install -e ".[dev]"
#
a2wsgi==1.10.10
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-asgi-connector==1.1.2
aiosignal==1.4.0
aiosqlite==0.22.0
alembic==1.17.2
altair==4.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.0
appdirs==1.4.4
asgiref==3.11.0
asyncer==0.0.8
attrs==25.4.0
backoff==2.2.1
bentoml==1.4.30
bigtree==1.0.4
black==25.12.0
blinker==1.9.0
boto3-stubs==1.42.13
botocore-stubs==1.42.13
build==1.3.0
cachetools==6.2.4
cattrs==23.1.2
causal-learn==0.1.4.3
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
clarabel==0.11.1
click==8.3.1
click-option-group==0.5.9
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
coverage==7.13.0
cryptography==46.0.3
cvxpy==1.7.5
cycler==0.12.1
Cython==3.2.3
dask==2025.12.0
databricks-sdk==0.76.0
dataclasses-json==0.6.7
datasets==4.4.2
deprecation==2.1.0
dill==0.3.9
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
dowhy==0.14
dspy==3.0.4
dspy-ai==3.0.4
econml==0.16.0
entrypoints==0.4
FalkorDB==1.2.2
fastapi==0.125.0
fasttext==0.9.3
fastuuid==0.14.0
feast==0.58.0
filelock==3.20.1
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
gepa==0.0.17
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.45.0
graphene==3.4.3
graphiti-core==0.24.3
graphql-core==3.2.7
graphql-relay==3.2.0
graphviz==0.21
great_expectations==1.10.0
greenlet==3.3.0
grpcio==1.67.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
httpx-ws==0.8.2
huey==2.5.5
huggingface-hub==0.36.0
hyperframe==6.1.0
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
instructor==1.13.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.3
json_repair==0.54.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kantoku==0.18.3
kiwisolver==1.4.9
langchain==1.2.0
langchain-anthropic==1.3.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.3
langchain-openai==1.1.6
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.0
librt==0.7.4
lightgbm==4.6.0
litellm==1.80.10
llvmlite==0.46.0
locket==1.0.0
magicattr==0.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.8
mdurl==0.1.2
mistune==3.1.4
mlflow==3.7.0
mlflow-skinny==3.7.0
mlflow-tracing==3.7.0
mmh3==5.2.0
momentchi2==0.1.8
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
mypy==1.19.1
mypy-boto3-bedrock-runtime==1.42.3
mypy_extensions==1.1.0
neo4j==6.0.3
nest-asyncio==1.6.0
networkx==3.6.1
nodeenv==1.9.1
numba==0.63.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.590.44
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==2.14.0
opentelemetry-api==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-aiohttp-client==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
opik==1.9.60
optuna==4.6.0
orjson==3.11.5
ormsgpack==1.12.1
osqp==1.0.5
packaging==25.0
pandas==2.3.3
partd==1.4.2
pathspec==0.12.1
patsy==1.0.2
pgvector==0.4.2
pillow==12.0.0
pip-requirements-parser==32.0.1
pip-tools==7.5.2
platformdirs==4.5.1
pluggy==1.6.0
postgrest==2.27.0
posthog==7.4.0
pre_commit==4.5.1
prometheus_client==0.23.1
propcache==0.4.1
protobuf==6.33.2
psutil==7.1.3
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==3.0.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydot==4.0.1
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.2.5
pyproject_hooks==1.2.0
pyroaring==1.0.3
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
ragas==0.4.1
RapidFuzz==3.14.3
realtime==2.27.0
redis==6.4.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.17.1
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
ruff==0.14.10
safetensors==0.7.0
schema==0.7.8
scikit-learn==1.6.1
scikit-network==0.33.5
scipy==1.15.3
scs==3.2.9
sentence-transformers==5.2.0
sentry-sdk==2.48.0
setuptools==80.9.0
shap==0.48.0
shellingham==1.5.4
simple-di==0.1.5
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sparse==0.17.0
SQLAlchemy==2.0.45
sqlparse==0.5.5
starlette==0.50.0
statsmodels==0.14.6
storage3==2.27.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.0
supabase-auth==2.27.0
supabase-functions==2.27.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tomli_w==1.2.0
toolz==1.1.0
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
ty==0.0.4
typeguard==4.4.4
typer==0.20.1
types-awscrt==0.30.0
types-s3transfer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uuid6==2025.0.1
uuid_utils==0.12.0
uvicorn==0.34.0
uvicorn-worker==0.3.0
uvloop==0.22.1
virtualenv==20.35.4
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.4
wheel==0.45.1
wrapt==1.17.3
wsproto==1.3.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
